Change the type sig on write_fixed to bytes
This was the source of the mysterious bug where the datafile test
would fail. Since the sync marker is 16 random bytes, they could
contain null markers (terminating the char* it was before) and
truncating the data to a smaller write. This caused the reader to
get out of sync since it was expecting 16 fixed bytes. The random
sync marker was what was making the test non-deterministic which
was frustrating.