Fix install from sdist when Cython is installed
Installing from the PyPI sdist failed when one has Cython installed.
This is because the sdist did not include the pyx file in the
source distribution. The setup.py tested that Cython was installed,
and when it was, it tried to compile from the pyx source, even though
it was not included in the sdist.

Fix inclusion of the pyx file, CHANGELOG.md, LICENSE.txt, NOTICE.txt
in the source distribution.