Re-generate fast_binary.c
Python 3.9 removed the function `tp_print`, causing the cython extension
to fail to compile. This simply re-generates the extension on python 3.9
with a newer version of cython.

I have no idea if there are other ramifications here or if anything needs
to be tested on older versions -- let me know if so.